Added Task, Note and TaskType Services and controller
