+++ entity-processor/src/main/java/com/fincity/saas/entity/processor/dao/base/BaseProcessorDAO.java
@@ -9,7 +9,7 @@ import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 public abstract class BaseProcessorDAO<R extends UpdatableRecord<R>, D extends BaseProcessorDto<D>>
-        extends BaseDAO<R, D> {
+        extends BaseUpdatableDAO<R, D> {
 
     protected BaseProcessorDAO(Class<D> flowPojoClass, Table<R> flowTable, Field<ULong> flowTableId) {
         super(flowPojoClass, flowTable, flowTableId);
