+++ entity-processor/src/main/java/com/fincity/saas/entity/processor/dao/rule/SimpleComplexRuleRelationDAO.java
@@ -2,7 +2,7 @@ package com.fincity.saas.entity.processor.dao.rule;
 
 import static com.fincity.saas.entity.processor.jooq.Tables.ENTITY_PROCESSOR_SIMPLE_COMPLEX_RULE_RELATIONS;
 
-import com.fincity.saas.entity.processor.dao.base.BaseDAO;
+import com.fincity.saas.entity.processor.dao.base.BaseUpdatableDAO;
 import com.fincity.saas.entity.processor.dto.rule.SimpleComplexRuleRelation;
 import com.fincity.saas.entity.processor.jooq.tables.records.EntityProcessorSimpleComplexRuleRelationsRecord;
 import org.jooq.Field;
@@ -13,7 +13,7 @@ import reactor.core.publisher.Mono;
 
 @Component
 public class SimpleComplexRuleRelationDAO
-        extends BaseDAO<EntityProcessorSimpleComplexRuleRelationsRecord, SimpleComplexRuleRelation> {
+        extends BaseUpdatableDAO<EntityProcessorSimpleComplexRuleRelationsRecord, SimpleComplexRuleRelation> {
 
     private final Field<ULong> complexConditionIdField;
