+++ entity-processor/src/main/java/com/fincity/saas/entity/processor/dao/rule/base/BaseRuleDAO.java
@@ -1,6 +1,6 @@
 package com.fincity.saas.entity.processor.dao.rule.base;
 
-import com.fincity.saas.entity.processor.dao.base.BaseDAO;
+import com.fincity.saas.entity.processor.dao.base.BaseUpdatableDAO;
 import com.fincity.saas.entity.processor.dto.rule.base.BaseRule;
 import com.fincity.saas.entity.processor.enums.EntitySeries;
 import org.jooq.Condition;
@@ -9,7 +9,7 @@ import org.jooq.Table;
 import org.jooq.UpdatableRecord;
 import org.jooq.types.ULong;
 
-public abstract class BaseRuleDAO<R extends UpdatableRecord<R>, D extends BaseRule<D>> extends BaseDAO<R, D> {
+public abstract class BaseRuleDAO<R extends UpdatableRecord<R>, D extends BaseRule<D>> extends BaseUpdatableDAO<R, D> {
 
     private static final String PRODUCT_STAGE_RULE_ID = "PRODUCT_STAGE_RULE_ID";
     private static final String PRODUCT_TEMPLATE_RULE_ID = "PRODUCT_TEMPLATE_RULE_ID";
