+++ entity-processor/src/main/java/com/fincity/saas/entity/processor/dto/Activity.java
@@ -1,0 +1,15 @@
+package com.fincity.saas.entity.processor.dto;
+
+import com.fincity.saas.entity.processor.dto.base.BaseUpdatableDto;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.ToString;
+import lombok.experimental.Accessors;
+import lombok.experimental.FieldNameConstants;
+
+@Data
+@Accessors(chain = true)
+@EqualsAndHashCode(callSuper = true)
+@ToString(callSuper = true)
+@FieldNameConstants
+public class Activity extends BaseUpdatableDto<Activity> {}
