+++ entity-processor/src/main/java/com/fincity/saas/entity/processor/dto/base/BaseProcessorDto.java
@@ -14,7 +14,8 @@ import org.springframework.data.annotation.Version;
 @EqualsAndHashCode(callSuper = true)
 @ToString(callSuper = true)
 @FieldNameConstants
-public abstract class BaseProcessorDto<T extends BaseProcessorDto<T>> extends BaseDto<T> implements IEntitySeries {
+public abstract class BaseProcessorDto<T extends BaseProcessorDto<T>> extends BaseUpdatableDto<T>
+        implements IEntitySeries {
 
     @Serial
     private static final long serialVersionUID = 5174424228629814984L;
