+++ entity-processor/src/main/java/com/fincity/saas/entity/processor/dto/base/BaseUpdatableDto.java
@@ -1,9 +1,15 @@
 package com.fincity.saas.entity.processor.dto.base;
 
-import com.fincity.saas.commons.jooq.flow.dto.AbstractFlowDTO;
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fincity.saas.commons.jooq.flow.dto.AbstractFlowUpdatableDTO;
 import com.fincity.saas.commons.util.UniqueUtil;
 import com.fincity.saas.entity.processor.model.base.BaseResponse;
+import com.fincity.saas.entity.processor.util.IClassConvertor;
 import java.io.Serial;
+import java.util.List;
+import java.util.Map;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
@@ -16,17 +22,28 @@ import org.jooq.types.ULong;
 @EqualsAndHashCode(callSuper = true)
 @ToString(callSuper = true)
 @FieldNameConstants
-public abstract class BaseDto<T extends BaseDto<T>> extends AbstractFlowDTO<ULong, ULong> {
+public class BaseUpdatableDto<T extends BaseUpdatableDto<T>> extends AbstractFlowUpdatableDTO<ULong, ULong>
+        implements IClassConvertor {
 
     public static final int CODE_LENGTH = 22;
 
     @Serial
-    private static final long serialVersionUID = 717874495505090612L;
+    private static final long serialVersionUID = 1844345864104376760L;
 
     private String code = UniqueUtil.shortUUID();
 
-    private String name;
+    private String name = this.code;
     private String description;
+
+    @JsonIgnore
+    private boolean tempActive = Boolean.FALSE;
+
+    private boolean isActive = Boolean.TRUE;
+
+    public static <T extends BaseUpdatableDto<T>> Map<ULong, T> toIdMap(List<T> baseDtoList) {
+        return baseDtoList.stream()
+                .collect(Collectors.toMap(BaseUpdatableDto::getId, Function.identity(), (a, b) -> b));
+    }
 
     public T setCode() {
         if (this.code != null) return (T) this;
@@ -43,6 +60,16 @@ public abstract class BaseDto<T extends BaseDto<T>> extends AbstractFlowDTO<ULon
     public T setDescription(String description) {
         if (description == null || description.isBlank()) return (T) this;
         this.description = description;
+        return (T) this;
+    }
+
+    public T setActive(boolean isActive) {
+        this.isActive = isActive;
+        return (T) this;
+    }
+
+    public T setTempActive(boolean tempActive) {
+        this.tempActive = tempActive;
         return (T) this;
     }
