+++ entity-processor/src/main/java/com/fincity/saas/entity/processor/dto/content/TaskType.java
@@ -1,8 +1,9 @@
 package com.fincity.saas.entity.processor.dto.content;
 
-import com.fincity.saas.entity.processor.dto.base.BaseDto;
+import com.fincity.saas.entity.processor.dto.base.BaseUpdatableDto;
 import com.fincity.saas.entity.processor.enums.EntitySeries;
 import com.fincity.saas.entity.processor.enums.IEntitySeries;
+import com.fincity.saas.entity.processor.model.request.content.TaskTypeRequest;
 import java.io.Serial;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
@@ -15,10 +16,14 @@ import lombok.experimental.FieldNameConstants;
 @EqualsAndHashCode(callSuper = true)
 @ToString(callSuper = true)
 @FieldNameConstants
-public class TaskType extends BaseDto<TaskType> implements IEntitySeries {
+public class TaskType extends BaseUpdatableDto<TaskType> implements IEntitySeries {
 
     @Serial
     private static final long serialVersionUID = 3217238190792368556L;
+
+    public static TaskType of(TaskTypeRequest taskTypeRequest) {
+        return new TaskType().setName(taskTypeRequest.getName()).setDescription(taskTypeRequest.getDescription());
+    }
 
     @Override
     public EntitySeries getEntitySeries() {
