+++ entity-processor/src/main/java/com/fincity/saas/entity/processor/dto/content/base/BaseContentDto.java
@@ -1,6 +1,7 @@
 package com.fincity.saas.entity.processor.dto.content.base;
 
-import com.fincity.saas.entity.processor.dto.base.BaseDto;
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fincity.saas.entity.processor.dto.base.BaseUpdatableDto;
 import com.fincity.saas.entity.processor.enums.IEntitySeries;
 import com.fincity.saas.entity.processor.model.request.content.BaseContentRequest;
 import java.io.Serial;
@@ -17,8 +18,8 @@ import org.springframework.data.annotation.Version;
 @EqualsAndHashCode(callSuper = true)
 @ToString(callSuper = true)
 @FieldNameConstants
-public abstract class BaseContentDto<Q extends BaseContentRequest<Q>, T extends BaseContentDto<Q, T>> extends BaseDto<T>
-        implements IEntitySeries {
+public abstract class BaseContentDto<Q extends BaseContentRequest<Q>, T extends BaseContentDto<Q, T>>
+        extends BaseUpdatableDto<T> implements IEntitySeries {
 
     @Serial
     private static final long serialVersionUID = 5174424228629814984L;
@@ -33,10 +34,12 @@ public abstract class BaseContentDto<Q extends BaseContentRequest<Q>, T extends 
 
     public abstract T of(Q contentRequest);
 
+    @JsonIgnore
     public boolean isTicketContent() {
         return ticketId != null;
     }
 
+    @JsonIgnore
     public boolean isOwnerContent() {
         if (isTicketContent()) return Boolean.FALSE;
         return ownerId != null;
