+++ entity-processor/src/main/java/com/fincity/saas/entity/processor/dto/rule/Rule.java
@@ -1,6 +1,6 @@
 package com.fincity.saas.entity.processor.dto.rule;
 
-import com.fincity.saas.entity.processor.dto.base.BaseDto;
+import com.fincity.saas.entity.processor.dto.base.BaseUpdatableDto;
 import com.fincity.saas.entity.processor.enums.EntitySeries;
 import com.fincity.saas.entity.processor.enums.IEntitySeries;
 import com.fincity.saas.entity.processor.enums.rule.DistributionType;
@@ -23,7 +23,7 @@ import org.springframework.data.annotation.Version;
 @EqualsAndHashCode(callSuper = true)
 @ToString(callSuper = true)
 @FieldNameConstants
-public abstract class Rule<T extends Rule<T>> extends BaseDto<T> implements IEntitySeries {
+public abstract class Rule<T extends Rule<T>> extends BaseUpdatableDto<T> implements IEntitySeries {
 
     public static final Map<String, Table<?>> relationsMap = Map.of(Fields.stageId, EntitySeries.STAGE.getTable());
