+++ entity-processor/src/main/java/com/fincity/saas/entity/processor/model/common/Identity.java
@@ -11,7 +11,7 @@ import com.fasterxml.jackson.databind.SerializerProvider;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.fincity.saas.commons.jooq.util.ULongUtil;
-import com.fincity.saas.entity.processor.dto.base.BaseDto;
+import com.fincity.saas.entity.processor.dto.base.BaseUpdatableDto;
 import java.io.IOException;
 import java.io.Serial;
 import java.io.Serializable;
@@ -75,7 +75,7 @@ public class Identity implements Serializable {
     public Identity setCode(String code) {
 
         if (code == null) return this;
-        if (code.length() == BaseDto.CODE_LENGTH) this.code = code;
+        if (code.length() == BaseUpdatableDto.CODE_LENGTH) this.code = code;
 
         try {
             this.id = new BigInteger(code);
